refactor: extract Expose.Params and Expose.Handlers from expose.ex
Move ~200 lines of param generation logic (generate_params with 10 clauses,
filter suffixes, type mapping helpers) into Ectomancer.Expose.Params.

Move ~130 lines of handler generation (select_handler, generate_simple_handler,
generate_upsert_handler, generate_batch_handler, preload handlers, field_auth
wrapper) into Ectomancer.Expose.Handlers.

Make get_ecto_type_for_param/1 public so both new modules can use it.

Reduce expose.ex from 1144 to 735 lines.
